requirements.txt
streamlit
python-dotenv
SQLAlchemy
pandas
pymysql
openai
python-jobspy
fpdf2
requests
setuptools
